adding generic functionality

+++ codingtest/src/main/java/codingtest/CheckTriangle.java
@@ -19,7 +19,7 @@ public class CheckTriangle {
 	 * @return true if they are valid lengths. Otherwise False
 	 */
 
-	private boolean validLength(int sideA, int sideB, int sideC) {
+	private boolean validLength(double sideA, double sideB, double sideC) {
 		return sideA > 0 && sideB > 0 && sideC > 0;
 	}
 
@@ -31,7 +31,7 @@ public class CheckTriangle {
 	 * @param sideC length of sideC in a triangle
 	 * @return true if the lengths does not adhere to triangle property. Otherwise return false.
 	 */
-	private boolean checkTraingleProperty(int sideA, int sideB, int sideC) {
+	private boolean checkTraingleProperty(double sideA, double sideB, double sideC) {
 		return sideA > sideB + sideC || sideB > sideA + sideC || sideC > sideA + sideB;
 	}
 	
@@ -44,7 +44,7 @@ public class CheckTriangle {
 	 * @return true if any two side's lengths are same.
 	 */
 
-	private boolean checkIsosceles(int sideA, int sideB, int sideC) {
+	private boolean checkIsosceles(double sideA, double sideB, double sideC) {
 		return sideA == sideB || sideB == sideC || sideA == sideC;
 	}
 	
@@ -57,28 +57,28 @@ public class CheckTriangle {
 	 * @return true if all sides lengths are same.
 	 */
 
-	private boolean checkEquilateral(int sideA, int sideB, int sideC) {
+	private boolean checkEquilateral(double sideA, double sideB, double sideC) {
 		return sideA == sideB && sideB == sideC;
 	}
 	
 	/**
 	 * This function returns the type of triangle if its valid. Else reason out the problem.
 	 * 
-	 * @param sideA, Integer length of sideA in a triangle
-	 * @param sideB, Integer length of sideB in a triangle
-	 * @param sideC, Integer length of sideC in a triangle
+	 * @param sideA, Double length of sideA in a triangle
+	 * @param sideB, Double length of sideB in a triangle
+	 * @param sideC, Double length of sideC in a triangle
 	 * @return the type of the triangle if the sides passed are valid, else reason out the problem
 	 */
 
-	public String getTriangleType(Integer sideA, Integer sideB, Integer sideC) {
+	public String getTriangleType(Double sideA, Double sideB, Double sideC) {
 		
 		if ((sideA != null && sideB != null && sideC != null)
-				&& validLength(sideA.intValue(), sideB.intValue(), sideC.intValue())
-				&& !checkTraingleProperty(sideA.intValue(), sideB.intValue(), sideC.intValue())) {
+				&& validLength(sideA.doubleValue(), sideB.doubleValue(), sideC.doubleValue())
+				&& !checkTraingleProperty(sideA.doubleValue(), sideB.doubleValue(), sideC.doubleValue())) {
 			
-			if (checkEquilateral(sideA.intValue(), sideB.intValue(), sideC.intValue())) {
+			if (checkEquilateral(sideA.doubleValue(), sideB.doubleValue(), sideC.doubleValue())) {
 				return "Its an Equilateral Triangle";
-			} else if (checkIsosceles(sideA.intValue(), sideB.intValue(), sideC.intValue())) {
+			} else if (checkIsosceles(sideA.doubleValue(), sideB.doubleValue(), sideC.doubleValue())) {
 				return "Its an Isoceles Triangle";
 			} else {
 				return "Its a Scalene Triangle";
